fix: follow ASDF's new way of binding *PACKAGE*
the issue is explained in:
https://gitlab.common-lisp.net/asdf/asdf/-/issues/157

basically ASDF started to bind *PACKAGE* (one of the numerous global
state) and it broke our way of loading/installing our reader macros.

this fix is more of a trial and error result than a well understood
change. e.g. both the :AROUND and the AROUND-COMPILE-HOOK are needed,
and i don't really know why the latter is also needed.